requirements.txt

requests>=2.31.0
PyYAML>=6.0.1
PyPDF2>=3.0.1
rapidfuzz>=3.9.7
